[FIX] product_pricelist_direct_print: empty categ. in last ordered products
When categories are set on the filter and and last ordered products is
checked, if one of the categories was empty it causes an error due to
the t-value returning an empty dict instead of an empty recordset.

The fix returns an empty recordset of either product.template o
product.product model.