[ADD] product_order_noname: Speedup product retrieve
This module improves retrieving products when a language is installed
and configured. The default behaviour is to order products by name,
which in this case would be slow because Odoo will process the
product's name to translate it.

When ordering products by name and having a language configured the
Odoo's ORM makes a join to `product_template` an then another join but
this time with a nested SELECT to retrieve the translation. This is very
slow because the lack of index on the JOIN, making a query like
`env['product.product'].search([], limit=10)` taking up to 1027.282 ms.

With the changes introduced here the query is considerably simpler, just
as the following:

    SELECT "product_product".id
    FROM   "product_product"
    WHERE  ( "product_product"."active" = true )
    ORDER  BY "product_product"."default_code"
    LIMIT  10

Also the execution time dramatically decrease, from 1027.282 ms to
0.624 ms which is about 1.65k times faster.