Fix barcode required error on variants creation
When we create a product template and select variant attributes, it
creates variants under the hood. The context key in
`_create_variant_ids` allows to bypass the check on the barcode, which
would be blocking.

If we have a related or computed field triggered by a create/write on a
product.template (`@api.depends('product_tmpl_id.foo')`), the check is
triggered and blocks the creation/write on the template.

We have then to ignore the check on variants on any create/write on
templates.