[FIX] product_pricelist_supplierinfo: Avoid min qty in Supplierinfo
Before this commit, using the field `no_supplierinfo_min_quantity` in the
`product.pricelist.item` only allow us to use `product.supplierinfo`
with `min_qty` equals 0 or 1, even when there is a supplierinfo with the
same sequence but a higher `min_qty` and the order of that model is
'sequence, min_qty desc, price' which tell us that the seller with a
higher `min_qty` should be first.

For example in the test `test_pricelist_min_quantity` when the field
`no_supplierinfo_min_quantity` is set to True the order of the
supplierinfo arrive first the one with the vendor "cls.supplier1" and
the price 50 because both sellers have the same sequence but this has
the higher `min_qty, but with the method `_select_seller` sending the
qty with 1 only allows us to use the first supplierinfo with the qty of
1 or 0 so that is why the price that was obtained was 10 instead of 50.

Adding a param to the method `_select_seller` in the model
`product.product` 'avoid_min_qty' that can be sent in the parameter
`params` where we will avoid the validation of the `min_qty` of the
supplierinfo to use on the method `_get_supplierinfo_pricelist_price` of
the model `product.template` when the field `no_supplierinfo_min_quantity`
is set as True.