[IMP] product_state (Move Data to Demo)
The stages in the data file is demo data. Having it in data causes issues when the module is updated and the sample stages were deleted as it will repopulate them, even though there is the no update flag.

It was already stated in the roadmap readme for v15, but why not fix it here.