Allow removal and reuse of existing sequence on product category
If a sequence prefix is defined on the product category, there's
no way through the UI to remove it. If the prefix field is emptied,
the existing sequence will still be defined on the category without
any possibility for the user to change the setting.
With this change, removing the prefix will remove the link between
the product category and the sequence.
Moreover, if an existing prefix is redefined on a product category
a new sequence was created instead of reusing an existing one.