Allow edition of product category sequence with constraints
As the product.category.sequence_id field of product.category is
only displayed in debug mode, we can allow its edition to ease
fixing of wrong configuration.
However, we need to make sure the prefix defined on these product
categories always matches the prefix defined on the sequence and
restrict any change that would introduce a discrepancy.