prod_lot_seq: Fix sequence incrementation opening detailed operations
When opening the detailed operations view of a stock move, Odoo is calling
stock.production.lot._get_next_serial to set stock.move.next_serial field
anytime the product is tracked by serial and the move is assigned.

If we use this module with product or global policy, the respective sequence
will therefore be called and incremented anytime this view is open, even
if the picking is not creating lots (e.g. delivery orders or internal transfers)

To avoid incrementing the sequence unnecessarily, we only allow to get the next
sequence number the first time this view is opened and if the move has
to create new serial numbers.
Otherwise, we force the value to be set to stock.move.next_serial field to
an empty string if no serial has to be created through this move, or to the
next_serial value assigned on the first opening of the view.