Add technical field `show_percent_change_button`
When buttons are clicked on PricelistItem pseudo-record
:meth:`create()` will be executed and new record will be
flushed in DB. This is a very dangerous behaviour because
rule is flushed unintentionally.

To bypass this issue:

we use a technical field `show_percent_change_button` that we
will set True on create(), and show button only if field is True.

This still pretty dangerous in case of:
* call from external API
* attrs in removed from XML header definition for some reason

In fact, it would be MUCH better to check field value python side
and raise UserError when technical field is false after create,
instead of manipulating button visibility in XML side. On the other
hand the situation is a little bit more complex to manage by using
:exc:`~odoo.exceptions.UserError` since :meth:`set_percentage_change()`
will be called after :meth:`create()` and python context is not shared
between two methods, this basically makes technical field unreliable
at python-side.