Improve fields_view_get() pricelist item implementation
To open the primary view when rule is opened by product
we use a special context-key in `fixed_pricelist_item_ids`
one2many field. There are very rare case when this might
be an issue: when you open a rule by products and want
open other dialogs (wizards) if the one2many rule dialog
is still opened in background, the context-key is still
active in the context. At this point if you try to return
another view for the pricelist-rule dialog, the renderer
will be forced to this primary product item view. It is
an extremely rare case, in such a scenario you can add
this `skip_primary_item_view_render` to the context.
It will bypass the precence of `open_pricelist_item_by_
product` key in `fixed_pricelist_item_ids` context.