Optimize price computation
Premise:

I faced performance issues due to the fact that when we fetch any
data from `item_ids`, on `product.pricelist` model, when you access
the whole recordset (e.g. api.depends or api.constrains..) you will
also make a fetch for field `product_pricelist_selling_price`.

This is an issue on large recordset of rules based on product, because
the function to calculate price might be havy to compute, expecially
if rules are based on "other pricelists". Note that you generally fetch
only a few records per-time on view, this is mostly a prefetch issue.

In order to avoid this issue I added a context key `compute_rule_selling_prices`
which must be detected as 'True' on context to actually trigger the full price
computation.

The issue with this workaround is that when field `product_pricelist_selling_price`
is non-stored, is not actually possible to simply skip recomputation by using :meth:
`self.env.remove_to_compute`: it would raise a `cacheMiss KeyError`. It seems that
you actually have to loop the whole recordset and provide a 'dummy value'.
The latter solution of providing 'dummy value' will avoid `cacheMiss KeyError` and
will definitely be better than running full computation when the fetch `product_
pricelist_selling_price` field is not requested, but it's not an optimal solution
since we still have O(n) computation instead of O(1) every time we access field
prefetch, even if the computation should not be necessary.

To resume: I have tried to skip the prefetch for `product_pricelist_selling_price`
field when recomputation is not requested explicitly by context, but I did not
managed to achieve it with non-stored (which I ended up to
set as default attribute for `product_pricelist_selling_price`).

---------------------------------------------------------------------------

Despite the 'non optimal state', the commit still ensure better performance
on large recordsets of pricelist rules by following steps:

a) adding the `compute_rule_selling_prices` evaluation on context to speed-up
   by making a 'dummy computation' when price visualization or fetch is not
   necessary (see the premise above). Now the module depends on module
   `base_view_inheritance_extension`, to extend context of one2many fields
   where key has to be placed, instead of overriding it.

b) removing `item_ids` from `_compute_product_pricelist_selling_price`
   dependencies. I think it's better in general to keep field unstored,
   to allow easy recomputation of multiple level of pricelist rules
   based on 'base_pricelist_id'. I also found out that `item_ids` dependency,
   which is not necessary when field is unstored, was causing massive performance
   issues on decent size recordsets.

c) a little boost is also obtained by removing method `_compute_base_pricelist_id
   _recursion_prefetch`, which wasn't really worth for general performance

Note: this commit also remove "optional" attribute on `item_ids` and `fixed_pricelist_
item_ids` tree view's fields. It wasn't in line with module purpose and has to be done
in a different module.
module, if necessary.