[IMP] product_supplierinfo_for_customer: better _select_customerinfo
Improve method `_select_customerinfo` to better mirror
`_select_seller` and use all its parameters

Some of the checks that were missing, compared to `select_seller`:

- didn't select a line according to start and end dates
- didn't select a line according to min qty
- didn't consider the parent contact, as well as the contact itself
- didn't order by lowest price
- didn't support UoM conversion

Now, the implementation as it relates to the SO, through the method
`compute_price` still doesn't use most of these. However, it does
benefit from some of these changes (for example, it will now look at
the parent contact, and it will take start and end dates into account).

My work also lays the foundation for future improvements,
and provides a good hook for other modules extending this one.
For example: https://github.com/OCA/stock-logistics-workflow/pull/1478
Here, the delivery address is often a child of the main SO contact,
and now it can make use of `_select_customerinfo` to select the correct
Customer Product Code and Customer Product Name.