[IMP] product_cost_security: clarify product cost edit required permissions
As seen in https://github.com/OCA/margin-analysis/pull/198#pullrequestreview-1951828542, a user needs some other group that grants them permissions for editing products, apart from this one that allows them to edit product cost.

This might not seem obvious to the user, so it's better to document it.

Btw, I renamed the groups to be shorter. Still easy to understand, now that groups are grouped (🥁).

@moduon MT-5158